Type fixups for libc::strlen()

+++ src/axfr-get.rs
@@ -886,7 +886,7 @@ pub unsafe extern "C" fn _c_main(mut argc: i32, mut argv: *mut *mut u8) -> i32 {
     if dns_domain_fromdot(
         &mut zone as (*mut *mut u8),
         *argv as (*const u8),
-        libc::strlen(*argv as (*const u8)),
+        libc::strlen(*argv as *const i8) as u32,
     ) == 0
     {
         die_generate();
